prisma-main/backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.0
pandas>=2.2.0
openpyxl==3.1.2 
python-dotenv>=1.0.0 